We don't want list styling in the event descriptions.
We allow for list in the event descriptions because we allow markdown.
We don't want to style it as such though because we are centering.  This
will never look good with centering.

Instead we get rid of the list styling but still encourage the user or list.
It will improve the look of the notification email.